+++ lib/src/service/model/training_data.dart
@@ -1,0 +1,13 @@
+class TrainingData {
+  final int trainingExamples;
+  final String baseModel;
+  final String language;
+  final double wordErrorRate;
+
+  const TrainingData({
+    required this.trainingExamples,
+    required this.baseModel,
+    required this.language,
+    required this.wordErrorRate,
+  });
+}