+++ lib/src/train/view/training_job_history_view.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
 import '../../common/error_indicator.dart';
 import '../../service/model/training_job.dart';
+import '../viewmodel/training_job_detail_viewmodel.dart';
 import '../viewmodel/training_job_history_viewmodel.dart';
+import 'training_job_detail_view.dart';
 
 class TrainingJobHistoryView extends StatefulWidget {
   const TrainingJobHistoryView({super.key, required this.viewModel});
@@ -20,6 +23,8 @@ class _TrainingJobHistoryViewState extends State<TrainingJobHistoryView> {
       builder: (context, orientation) {
         final double verticalPadding = 32;
         return Scaffold(
+          appBar:
+              AppBar(title: Text('Training job history'), centerTitle: true),
           body: ListenableBuilder(
             listenable: widget.viewModel.initializeModel,
             builder: (context, child) {
@@ -50,11 +55,6 @@ class _TrainingJobHistoryViewState extends State<TrainingJobHistoryView> {
               builder: (context, child) {
                 return CustomScrollView(
                   slivers: [
-                    SliverAppBar(
-                      title: Text('Training job history'),
-                      centerTitle: true,
-                      pinned: true,
-                    ),
                     SliverPadding(
                       padding: EdgeInsetsGeometry.symmetric(
                         vertical: verticalPadding,
@@ -128,7 +128,23 @@ class _TrainingJobHistoryViewState extends State<TrainingJobHistoryView> {
                                                   : Container()
                                             ])),
                               trailing: _trailingWidgetForListTile(
-                                  job, downloadStatus));
+                                  job, downloadStatus),
+                              onTap: _displayTrainingJobDetail(job.progress)
+                                  ? () {
+                                      Navigator.push(
+                                          context,
+                                          MaterialPageRoute(
+                                              builder: (context) =>
+                                                  TrainingJobDetailView(
+                                                      viewModel:
+                                                          TrainingJobDetailViewModel(
+                                                              modelRepository:
+                                                                  context
+                                                                      .read(),
+                                                              trainingId: job
+                                                                  .trainingId))));
+                                    }
+                                  : null);
                         }).toList(),
                       ),
                     ),
@@ -153,20 +169,30 @@ class _TrainingJobHistoryViewState extends State<TrainingJobHistoryView> {
     return Colors.redAccent;
   }
 
+  bool _displayTrainingJobDetail(String progress) {
+    return progress.toLowerCase() != 'failed';
+  }
+
   Widget _trailingWidgetForListTile(TrainingJob job, DownloadStatus status) {
     Widget? trailingWidget;
     switch (status) {
       case DownloadStatus.notStarted:
-        trailingWidget = MaterialButton(
-            onPressed: () {
-              widget.viewModel.downloadModel.execute(job.trainingId);
-            },
-            color: Colors.blueAccent,
-            textColor: Colors.white,
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.all(Radius.circular(80)),
-            ),
-            child: Text('Download'));
+        trailingWidget = Wrap(
+            crossAxisAlignment: WrapCrossAlignment.center,
+            spacing: 16,
+            children: [
+              MaterialButton(
+                  onPressed: () {
+                    widget.viewModel.downloadModel.execute(job.trainingId);
+                  },
+                  color: Colors.blueAccent,
+                  textColor: Colors.white,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(80)),
+                  ),
+                  child: Text('Download')),
+              Icon(Icons.chevron_right)
+            ]);
       case DownloadStatus.inProgress:
         trailingWidget = CircularProgressIndicator();
       case DownloadStatus.interrupted:
@@ -181,7 +207,13 @@ class _TrainingJobHistoryViewState extends State<TrainingJobHistoryView> {
             ),
             child: Text('Try again!'));
       case DownloadStatus.completed:
-        trailingWidget = Chip(label: Text('Available'));
+        trailingWidget = Wrap(
+            crossAxisAlignment: WrapCrossAlignment.center,
+            spacing: 16,
+            children: [
+              Chip(label: Text('Available')),
+              Icon(Icons.chevron_right)
+            ]);
     }
     return Visibility(
         visible: job.progress == 'finished', child: trailingWidget);